requirements.txt
binance-futures-connector
termcolor
python-telegram-bot[job-queue]
